+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.2.0] / 2022-07-09
-### Updated
+### Added
 - Add `CreatePushButton` and Obsolete `AddPushButton`
+### Updated
+- Update `RibbonSettings` Add with Action
+- Update `RibbonDescription` (LanguageType) 
 
 ## [0.1.2] / 2022-05-17
 ### Added